refactor/스크린 폴더 구조 변경
