+++ src/screens/Login/NicknameWrite/index.tsx
@@ -88,7 +88,7 @@ const NicknameWriteScreen = ({navigation}: Readonly<AuthProps>) => {
             />
           </View>
           <Image
-            source={require('../../../assets/pngs/background/background2.png')}
+            source={require('../../../../assets/pngs/background/background2.png')}
             className="w-full h-auto flex-1 mt-[79]"
           />
           <View className="absolute left-0 bottom-[30] w-full px-[40]">
