+++ src/screens/Login/VolunteerOnboarding/index.tsx
@@ -69,7 +69,7 @@ const Page2 = () => {
         className="text-gray200 text-center"
       />
       <Image
-        source={require('../../../assets/pngs/background/background3.png')}
+        source={require('../../../../assets/pngs/background/background3.png')}
         className="w-full h-[466] absolute bottom-0"
       />
     </View>
@@ -87,7 +87,7 @@ const Page3 = ({nickname}: Readonly<PageProps>) => {
         className="text-gray200 text-center"
       />
       <Image
-        source={require('../../../assets/pngs/background/background1.png')}
+        source={require('../../../../assets/pngs/background/background1.png')}
         className="w-full h-auto"
       />
     </View>
@@ -105,7 +105,7 @@ const Page4 = ({handleNext}: Readonly<{handleNext: () => void}>) => {
         className="text-gray200 text-center "
       />
       <Image
-        source={require('../../../assets/pngs/constellation.png')}
+        source={require('../../../../assets/pngs/constellation.png')}
         width={274}
         height={269.5}
         className="w-[274] h-[269.5] mt-[100]"
