+++ src/screens/RCD/RCDFeedBack/index.tsx
@@ -68,7 +68,7 @@ const RCDFeedBackScreen = () => {
             <View className="w-full h-full justify-center items-center">
               <View className="w-[90%] h-[85%] justify-center items-center">
                 <ImageBackground
-                  source={require('../../../assets/pngs/starsOff.png')}
+                  source={require('../../../../assets/pngs/starsOff.png')}
                   resizeMode="contain"
                   className="static w-full h-full"
                 />
@@ -83,7 +83,7 @@ const RCDFeedBackScreen = () => {
               <View className="w-full h-full justify-center items-center">
                 <View className="w-[90%] h-[85%] justify-center items-center">
                   <ImageBackground
-                    source={require('../../../assets/pngs/starsOn.png')}
+                    source={require('../../../../assets/pngs/starsOn.png')}
                     resizeMode="contain"
                     className="w-full h-full"
                   />
